# libs/api-client/src/generated/model/createOrganizationDto.ts
/**
 * Portfolio API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface CreateOrganizationDto {
  name: string;
  logoUrl?: string;
  /**
   * R2 object key returned by the logo presign endpoint, paired with `logoUrl`
   */
  logoObjectKey?: string;
  website?: string;
}