# openapi/api.yaml
openapi: 3.0.0
paths:
...
  /api/projects/upload-url:
    post:
      operationId: ProjectsController_createImageUploadUrl
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectImageUploadUrlDto'
      responses:
        '200':
          description: A short-lived presigned PUT url for a project image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectImageUploadUrlDto'
        '401':
          description: No valid session
      tags: &ref_2
        - projects
  /api/projects:
    get:
      operationId: ProjectsController_findAll
      parameters: []
      responses:
        '200':
          description: All projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectDto'
      tags: *ref_2
    post:
      operationId: ProjectsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectDto'
      responses:
        '201':
          description: The created project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDto'
        '401':
          description: No valid session
        '409':
          description: A project with this slug already exists
      tags: *ref_2
  /api/projects/{id}/related:
    get:
      operationId: ProjectsController_findRelated
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: limit
          required: false
          in: query
          description: Max related projects to return (default 3)
          schema:
            type: number
      responses:
        '200':
          description: Other projects related by shared skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectDto'
        '404':
          description: No project with this id exists
      tags: *ref_2
  /api/projects/{id}:
    get:
      operationId: ProjectsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: The requested project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDto'
        '404':
          description: No project with this id exists
      tags: *ref_2
    put:
      operationId: ProjectsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectDto'
      responses:
        '200':
          description: The updated project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDto'
        '401':
          description: No valid session
        '404':
          description: No project with this id exists
        '409':
          description: A project with this slug already exists
      tags: *ref_2
    delete:
      operationId: ProjectsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The project was deleted
        '401':
          description: No valid session
        '404':
          description: No project with this id exists
      tags: *ref_2
...
  /api/organizations/upload-url:
    post:
      operationId: OrganizationsController_createLogoUploadUrl
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationLogoUploadUrlDto'
      responses:
        '200':
          description: A short-lived presigned PUT url for an organization logo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationLogoUploadUrlDto'
        '401':
          description: No valid session
      tags: &ref_3
        - organizations
  /api/organizations:
    get:
      operationId: OrganizationsController_findAll
      parameters: []
      responses:
        '200':
          description: All organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrganizationDto'
      tags: *ref_3
    post:
      operationId: OrganizationsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationDto'
      responses:
        '201':
          description: The created organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationDto'
        '401':
          description: No valid session
        '409':
          description: An organization with this name already exists
      tags: *ref_3
  /api/organizations/{id}:
    get:
      operationId: OrganizationsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: The requested organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationDto'
        '404':
          description: No organization with this id exists
      tags: *ref_3
    put:
      operationId: OrganizationsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrganizationDto'
      responses:
        '200':
          description: The updated organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationDto'
        '401':
          description: No valid session
        '404':
          description: No organization with this id exists
        '409':
          description: An organization with this name already exists
      tags: *ref_3
    delete:
      operationId: OrganizationsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The organization was deleted
        '401':
          description: No valid session
        '404':
          description: No organization with this id exists
        '409':
          description: The organization still has roles referencing it
      tags: *ref_3
...
  /api/resume/upload-url:
    post:
      operationId: ResumeController_createUploadUrl
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResumeUploadUrlDto'
      responses:
        '200':
          description: A short-lived presigned PUT url
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeUploadUrlDto'
        '401':
          description: No valid session
      tags: &ref_7
        - resume
  /api/resume:
    put:
      operationId: ResumeController_confirmUpload
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmResumeUploadDto'
      responses:
        '200':
          description: The stored resume
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeDto'
        '401':
          description: No valid session
      tags: *ref_7
    get:
      operationId: ResumeController_findResume
      parameters: []
      responses:
        '200':
          description: The current resume metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeDto'
        '404':
          description: No resume uploaded yet
      tags: *ref_7
    delete:
      operationId: ResumeController_remove
      parameters: []
      responses:
        '204':
          description: The resume was deleted
        '401':
          description: No valid session
        '404':
          description: No resume uploaded yet
      tags: *ref_7
...
info:
  title: Portfolio API
  description: API documentation for the portfolio backend
  version: 0.0.1
  contact: {}
...
servers: []
components:
  schemas:
...
    CreateProjectImageUploadUrlDto:
      type: object
      properties:
        fileName:
          type: string
        mimeType:
          type: string
          enum:
            - image/png
            - image/jpeg
            - image/webp
        fileSize:
          type: number
          maximum: 5242880
      required:
        - fileName
        - mimeType
        - fileSize
    ProjectImageUploadUrlDto:
      type: object
      properties:
        uploadUrl:
          type: string
        objectKey:
          type: string
        publicUrl:
          type: string
          description: >-
            Public URL to store as the project imageUrl once the upload
            completes
      required:
        - uploadUrl
        - objectKey
        - publicUrl
    SkillDto:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the skill
          example: cly1x8f9z0000abcd1234efgh
        name:
          type: string
          example: Angular
      required:
        - id
        - name
    ProjectDto:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the project
          example: cly1x8f9z0000abcd1234efgh
        title:
          type: string
          example: Personal Portfolio
        slug:
          type: string
          description: URL-safe unique identifier
          example: personal-portfolio
        shortDescription:
          type: string
          description: Short summary shown alongside the project
          example: A fast, accessible portfolio built with Angular and NestJS.
          maxLength: 255
        description:
          type: string
          description: >-
            Markdown source for the project description, rendered client-side
            with ngx-markdown
          example: |-
            ## Overview

            Built with **Angular** and *NestJS*.
        imageUrl:
          type: string
          example: https://cdn.example.com/projects/portfolio.png
        client:
          type: string
          example: Acme Corp
        jobRole:
          type: string
          example: Lead Frontend Engineer
        liveUrl:
          type: string
          example: https://portfolio.example.com
        startDate:
          format: date-time
          type: string
          description: When work on the project started
          example: '2024-01-15T00:00:00.000Z'
        endDate:
          format: date-time
          type: string
          description: When work on the project ended — absent means still ongoing
          example: '2024-06-30T00:00:00.000Z'
        skills:
          description: Skills used on this project
          type: array
          items:
            $ref: '#/components/schemas/SkillDto'
        createdAt:
          format: date-time
          type: string
          example: '2026-08-14T12:00:00.000Z'
        updatedAt:
          format: date-time
          type: string
          example: '2026-08-14T12:00:00.000Z'
      required:
        - id
        - title
        - slug
        - shortDescription
        - description
        - startDate
        - skills
        - createdAt
        - updatedAt
    CreateProjectDto:
      type: object
      properties:
        title:
          type: string
          example: Personal Portfolio
        slug:
          type: string
          description: URL-safe unique identifier
          example: personal-portfolio
        shortDescription:
          type: string
          description: Short summary shown alongside the project
          example: A fast, accessible portfolio built with Angular and NestJS.
          maxLength: 255
        description:
          type: string
          description: >-
            Markdown source for the project description, rendered client-side
            with ngx-markdown
          example: |-
            ## Overview

            Built with **Angular** and *NestJS*.
        imageUrl:
          type: string
          example: https://cdn.example.com/projects/portfolio.png
        imageObjectKey:
          type: string
          description: >-
            R2 object key returned by the image presign endpoint, paired with
            `imageUrl`
          example: projects/9c3f9b2e-2b8e-4a9b-8f1a-9c3f9b2e2b8e.png
        client:
          type: string
          example: Acme Corp
        jobRole:
          type: string
          example: Lead Frontend Engineer
        liveUrl:
          type: string
          example: https://portfolio.example.com
        startDate:
          type: string
          description: When work on the project started
          example: '2024-01-15'
        endDate:
          type: string
          description: When work on the project ended — omit if still ongoing
          example: '2024-06-30'
        skills:
          description: >-
            Skill names used on this project (e.g. React, Angular, .NET, Figma)
            — created if they do not already exist
          example: &ref_8
            - Angular
            - NestJS
            - Figma
          type: array
          items:
            type: string
      required:
        - title
        - slug
        - shortDescription
        - description
        - startDate
    UpdateProjectDto:
      type: object
      properties:
        title:
          type: string
          example: Personal Portfolio
        slug:
          type: string
          description: URL-safe unique identifier
          example: personal-portfolio
        shortDescription:
          type: string
          description: Short summary shown alongside the project
          example: A fast, accessible portfolio built with Angular and NestJS.
          maxLength: 255
        description:
          type: string
          description: >-
            Markdown source for the project description, rendered client-side
            with ngx-markdown
          example: |-
            ## Overview

            Built with **Angular** and *NestJS*.
        imageUrl:
          type: string
          example: https://cdn.example.com/projects/portfolio.png
        imageObjectKey:
          type: string
          description: >-
            R2 object key returned by the image presign endpoint, paired with
            `imageUrl`
          example: projects/9c3f9b2e-2b8e-4a9b-8f1a-9c3f9b2e2b8e.png
        client:
          type: string
          example: Acme Corp
        jobRole:
          type: string
          example: Lead Frontend Engineer
        liveUrl:
          type: string
          example: https://portfolio.example.com
        startDate:
          type: string
          description: When work on the project started
          example: '2024-01-15'
        endDate:
          type: string
          description: When work on the project ended — omit if still ongoing
          example: '2024-06-30'
        skills:
          description: >-
            Skill names used on this project (e.g. React, Angular, .NET, Figma)
            — created if they do not already exist
          example: *ref_8
          type: array
          items:
            type: string
    CreateOrganizationLogoUploadUrlDto:
      type: object
      properties:
        fileName:
          type: string
        mimeType:
          type: string
          enum:
            - image/png
            - image/jpeg
            - image/webp
        fileSize:
          type: number
          maximum: 5242880
      required:
        - fileName
        - mimeType
        - fileSize
    OrganizationLogoUploadUrlDto:
      type: object
      properties:
        uploadUrl:
          type: string
        objectKey:
          type: string
        publicUrl:
          type: string
          description: >-
            Public URL to store as the organization logoUrl once the upload
            completes
      required:
        - uploadUrl
        - objectKey
        - publicUrl
    OrganizationDto:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the organization
          example: cly1x8f9z0000abcd1234efgh
        name:
          type: string
          example: Acme Corp
        logoUrl:
          type: string
          example: https://cdn.example.com/logos/acme.png
        website:
          type: string
          example: https://acme.example.com
        createdAt:
          format: date-time
          type: string
          example: '2026-08-17T12:00:00.000Z'
        updatedAt:
          format: date-time
          type: string
          example: '2026-08-17T12:00:00.000Z'
      required:
        - id
        - name
        - createdAt
        - updatedAt
    CreateOrganizationDto:
      type: object
      properties:
        name:
          type: string
          example: Acme Corp
        logoUrl:
          type: string
          example: https://cdn.example.com/logos/acme.png
        logoObjectKey:
          type: string
          description: >-
            R2 object key returned by the logo presign endpoint, paired with
            `logoUrl`
          example: organizations/9c3f9b2e-2b8e-4a9b-8f1a-9c3f9b2e2b8e.png
        website:
          type: string
          example: https://acme.example.com
      required:
        - name
    UpdateOrganizationDto:
      type: object
      properties:
        name:
          type: string
          example: Acme Corp
        logoUrl:
          type: string
          example: https://cdn.example.com/logos/acme.png
        logoObjectKey:
          type: string
          description: >-
            R2 object key returned by the logo presign endpoint, paired with
            `logoUrl`
          example: organizations/9c3f9b2e-2b8e-4a9b-8f1a-9c3f9b2e2b8e.png
        website:
          type: string
          example: https://acme.example.com
...
    CreateResumeUploadUrlDto:
      type: object
      properties:
        fileName:
          type: string
        mimeType:
          type: string
          enum:
            - application/pdf
        fileSize:
          type: number
          maximum: 5242880
      required:
        - fileName
        - mimeType
        - fileSize
    ResumeUploadUrlDto:
      type: object
      properties:
        uploadUrl:
          type: string
        objectKey:
          type: string
      required:
        - uploadUrl
        - objectKey
    ConfirmResumeUploadDto:
      type: object
      properties:
        objectKey:
          type: string
        fileName:
          type: string
        mimeType:
          type: string
          enum:
            - application/pdf
        fileSize:
          type: number
          maximum: 5242880
      required:
        - objectKey
        - fileName
        - mimeType
        - fileSize
    ResumeDto:
      type: object
      properties:
        fileName:
          type: string
        mimeType:
          type: string
        fileSize:
          type: number
        uploadedAt:
          format: date-time
          type: string
      required:
        - fileName
        - mimeType
        - fileSize
        - uploadedAt
...